Update README with Rails example
I know the Rails case is in the wiki, but it took me awhile to figure this out, and in the meantime I was stuck. I even posted a question on stackoverflow:

https://stackoverflow.com/questions/65224796/resque-not-failing-an-obviously-buggy-job

And one of the users suggested I request this change, as the Rails is pretty popular case for the Resque gem.