Prevent child process disconnecting parent connection to Redis
When a new task is started it is forked to run in its own process (to
protect the coordinator process from crashing if something goes wrong
in a task). In the child process the task reconnects to Redis (to avoid
sharing the connection with the parent process) but this reconnect is
implemented as a disconnect followed by a connect.

The disconnect here is not appropriate because the parent process might
be currently using that connection. When connecting to Redis using SSL
(which is becoming more common now Redis itself supports TLS natively)
this matters and causes the following error:

SSL_read: sslv3 alert bad record mac (OpenSSL::SSL::SSLError)

By calling "connect" here rather than "reconnect" we leave the parent
connection to Redis intact and create a new one for child processes if
they run with fork_per_job = true.